Fixed bug in spec <2> of reset() function

+++ ShiftAnd_unittest.cpp
@@ -213,10 +213,10 @@ TEST_F(ShiftAnd_test, matching_same)
 
     // query sequence
     std::vector<uint64_t> matchings0;
-    std::vector<uint16_t> errors0;
+    std::vector<uint8_t> errors0;
     sa0.querySeq(t.begin(), t.end(), matchings0, errors0);
     std::vector<uint64_t> matchings1;
-    std::vector<uint16_t> errors1;
+    std::vector<uint8_t> errors1;
     sa1.querySeq(t.begin(), t.end(), matchings1, errors1);
 
     // check the accepting masks
@@ -274,7 +274,7 @@ TEST_F(ShiftAnd_test, matching_smaller)
 
     // query the text to automata
     std::vector<uint64_t> matchings;
-    std::vector<uint16_t> errors;
+    std::vector<uint8_t> errors;
     sa1.querySeq(t.begin(), t.end(), matchings, errors);
 
 
@@ -317,7 +317,7 @@ TEST_F(ShiftAnd_test, matching_bisulfite)
     ShiftAnd<1> sa1(p, lmap);
 
     std::vector<uint64_t> matchings;
-    std::vector<uint16_t> errors;
+    std::vector<uint8_t> errors;
     sa1.querySeq(t.begin(), t.end(), matchings, errors);
 
     ASSERT_EQ(3, matchings.size());
@@ -344,3 +344,5 @@ TEST_F(ShiftAnd_test, matching_bisulfite)
     ASSERT_EQ(1, errors[0]);
 
 }
+
+// TODO long test?
